Criação de Service + Controllers

+++ src/main/java/com/sabinomecanica/backend/models/MovtoEntrada.java
@@ -22,7 +22,7 @@ public class MovtoEntrada {
 
     private Date data;
 
-    private float valor_pago;
+    private double valor_pago;
 
     public UUID getId() {
         return id;
@@ -48,11 +48,11 @@ public class MovtoEntrada {
         this.data = data;
     }
 
-    public float getValor_pago() {
+    public double getValor_pago() {
         return valor_pago;
     }
 
-    public void setValor_pago(float valor_pago) {
+    public void setValor_pago(double valor_pago) {
         this.valor_pago = valor_pago;
     }
 }
